requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
dnspython==2.6.1
eventlet==0.37.0
falcon==3.1.3
greenlet==3.1.0
gunicorn==23.0.0
idna==3.10
MarkupSafe==2.1.5
mongoengine==0.29.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pymongo==4.8.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
whitenoise==6.7.0
